requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
geopy==2.4.0
openai==1.3.5
phonenumbers==8.13.24
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
